requirements.txt
flask
pubchempy
retry
pandas
biopython
rdkit-pypi

